Mapping feature

+++ database/migrations/2019_11_04_101209_create_bike_stations_table.php
@@ -15,8 +15,8 @@ class CreateBikeStationsTable extends Migration
     {
         Schema::create('bike_stations', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->bigInteger('lat');
-            $table->bigInteger('long');
+            $table->float('lat');
+            $table->float('long');
             $table->timestamps();
         });
     }
